server/requirements.txt
asgiref==3.7.2
autopep8==2.3.1
Brotli==1.1.0
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
java-to-python-transpiler==0.1.8
packaging==24.0
pycodestyle==2.12.0
pytz==2023.4
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.28.0
whitenoise==6.6.0

